setCacheProducts() returns void + self::$cache_products is always array
I think the isset() call on setCacheProducts() is tricky, not readable
at all and should be avoided : This isset() return true for empty array
and false if defined to 'null', this implies to remember that, not so
common.

So I replaced it with is_null, really easier to understand (no
unexpected returns).
Null value is only possible if setCacheProducts() has never been called.
The rest of the time it will be an empty array.

Also I removed the return statement for better consistency.